requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.7.4
pycparser==2.19
python-telegram-bot==12.3.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
virtualenv==16.7.9
Werkzeug==0.16.0
yarl==1.4.2